Refactor to return Stream type
Previously, the subscribe method expected the user to pass a sending end
of a tokio channel which would be used to send updates to the
subscriber.

With this refactor, the subscribe method creates a pair of sender and
receiver for an unbound channel of the `futures` crate. It returns the
type `PubSubStream` which implements `Stream`.

The example binary and the function to create an echo client were
rewritten accordingly.